requirements.txt
Django==1.7.4
django-autoslug==1.7.2
mistune==0.5
wsgiref==0.1.2
